pyproject.toml
[tool.poetry]
name = "fastapi-shop"
version = "0.1.0"
description = ""
authors = ["dmitriy <dmitriydmgora@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"
alembic = "^1.13.2"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.32"}
python-slugify = "^8.0.4"
asyncpg = "^0.29.0"
dynaconf = "^3.2.6"
bcrypt = "^4.2.0"
pydantic = {extras = ["email"], version = "^2.8.2"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[tool.poetry.group.tests.dependencies]
mypy = "^1.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
